Replace Moran with fixed Moran dataset
